refactor: 3a (commands) — extract cmd_list + cmd_stack to their own modules
Move cmd_list to cli/cmd_list.py and cmd_stack_list/cmd_stack_apply to
cli/cmd_stack.py, promoting their previously-lazy imports to clean top-level
imports (cycle-free: none of marker/overlay/stack_select/setup_review/
skill_context import govkit at load time). govkit re-exports the 3 handlers for
main()'s dispatch and tests. Now-unused json + STACK_ID_ASSUMPTION imports
dropped. govkit.py: 974 -> 840 lines. No behavior change; 807 tests green.
